+++ client/widgets/markets/CAltarArtifacts.h
@@ -1,5 +1,5 @@
 /*
- * CAltar.h, part of VCMI engine
+ * CAltarArtifacts.h, part of VCMI engine
  *
  * Authors: listed in file AUTHORS in main folder
  *
@@ -9,10 +9,10 @@
  */
 #pragma once
 
-#include "../widgets/CArtifactsOfHeroAltar.h"
-#include "../widgets/CTradeBase.h"
+#include "../CArtifactsOfHeroAltar.h"
+#include "CTradeBase.h"
 
-class CAltarArtifacts : public CExpAltar
+class CAltarArtifacts : public CExperienceAltar
 {
 public:
 	CAltarArtifacts(const IMarket * market, const CGHeroInstance * hero);
@@ -43,29 +43,5 @@ private:
 	};
 
 	bool putArtOnAltar(std::shared_ptr<CTradeableItem> altarSlot, const CArtifactInstance * art);
-	void onSlotClickPressed(std::shared_ptr<CTradeableItem> altarSlot);
-};
-
-class CAltarCreatures : public CExpAltar, public CCreaturesSelling
-{
-public:
-	CAltarCreatures(const IMarket * market, const CGHeroInstance * hero);
-	void updateSlots();
-	void deselect() override;
-	TExpType calcExpAltarForHero() override;
-	void makeDeal() override;
-	void sacrificeAll() override;
-	void updateAltarSlot(std::shared_ptr<CTradeableItem> slot);
-
-private:
-	std::shared_ptr<CButton> maxUnits;
-	std::vector<int> unitsOnAltar;
-	std::vector<int> expPerUnit;
-	std::shared_ptr<CLabel> lSubtitle, rSubtitle;
-
-	void readExpValues();
-	void updateControls();
-	void updateSubtitlesForSelected();
-	void onOfferSliderMoved(int newVal);
-	void onSlotClickPressed(std::shared_ptr<CTradeableItem> & newSlot, std::shared_ptr<CTradeableItem> & hCurSide) override;
+	void onSlotClickPressed(const std::shared_ptr<CTradeableItem> & newSlot, std::shared_ptr<CTradeableItem> & hCurSlot) override;
 };
